+++ src/app/components/components.module.js
@@ -2,9 +2,11 @@
     'use strict';
 
     angular.module('chpl.components', [
+        'angularMoment',
         'angulartics',
         'chpl.services',
         'feature-flags',
+        'ngAvatar',
         'ngCytoscape',
         'ngFileUpload',
         'ngResource',