+++ RELEASE_NOTES.md
@@ -1,5 +1,24 @@
 # Release Notes
 
+## Version 10.34.0
+_20 December 2019_
+
+### Flagged Features
+* Create Change Request Administration view page
+
+---
+
+## Version 10.33.0
+_16 December 2019_
+
+### Features
+* Support custom parameters for recurring triggers
+
+### Flagged Features
+* Create Developer dashboard
+
+---
+
 ## Version 10.32.0
 _2 December 2019_